+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+	"eslint.workingDirectories": [
+		"src/ui/client"
+	]
+}
