+++ run-this-shit
@@ -1,0 +1,4 @@
+cd src/currencyexchange && yarn start &
+cd src/destination && yarn start &
+cd src/hotel && yarn start &
+cd src/ui && yarn start
