--- src/ui/client/src/components/BeeLogo.js
@@ -1,104 +1,0 @@
-import React from "react";
-
-const BeeLogo2 = ({ className, style }) => (
-  <svg viewBox="0 0 97 97" className={className} style={style}>
-    <defs>
-      <path
-        id="a"
-        d="M0 0.0119641761L4.05421204 0.0119641761 4.05421204 20.2370805 0 20.2370805z"
-      />
-    </defs>
-    <g fill="none" fill-rule="evenodd">
-      {/* <path
-        fill="#002236"
-        d="M48.5,1.88893345e-12 L48.5,1.88893345e-12 C75.2858104,1.88401298e-12 97,21.7141896 97,48.5 L97,48.5 C97,75.2858104 75.2858104,97 48.5,97 L48.5,97 C21.7141896,97 2.68802163e-12,75.2858104 2.68474132e-12,48.5 L2.68474132e-12,48.5 C2.681461e-12,21.7141896 21.7141896,1.89385393e-12 48.5,1.88893345e-12 Z"
-      /> */}
-      <path
-        fill="#E3B602"
-        d="M26.6914856 14.063463C28.508718 11.1306231 31.6027393 9.19789669 35.1177542 9.19789669 38.6323199 9.19789669 41.7266781 11.1306231 43.5440227 14.063463L26.6914856 14.063463zM25.0485109 18.9289351L45.1869301 18.9289351C45.2294785 19.3280104 45.2533909 19.7343612 45.2533909 20.1454277L45.2533909 23.7946361 24.98205 23.7946361 24.98205 20.1454277C24.98205 19.7343612 25.0059625 19.3280104 25.0485109 18.9289351zM24.9820725 32.3094109L24.9820725 28.6602025 45.2534134 28.6602025 45.2534134 32.3094109C45.2534134 32.7206121 45.2295009 33.1268282 45.1869525 33.5257688L25.0485333 33.5257688C25.0058727 33.1268282 24.9820725 32.7206121 24.9820725 32.3094109z"
-        transform="translate(13.99 27.048)"
-      />
-      <g transform="rotate(-90 55.064 14.808)">
-        <mask id="b" fill="#fff">
-          <path d="M0 0.0119641761L4.05421204 0.0119641761 4.05421204 20.2370805 0 20.2370805z" />
-        </mask>
-        <path
-          fill="#E3B602"
-          d="M-1.12264926e-05,8.90800036 C-1.12264926e-05,12.6185045 1.6105414,15.8854928 4.05421204,17.8040365 L4.05421204,0.0119641761 C1.6105414,1.93062644 -1.12264926e-05,5.19702213 -1.12264926e-05,8.90800036"
-          mask="url(#b)"
-        />
-      </g>
-      <path
-        fill="#27AA7C"
-        d="M17.8097271,17.73782 C17.3797982,16.7020026 16.7477718,15.7277576 15.9054291,14.8859539 C15.0618739,14.0423986 14.0905929,13.4117195 13.0542366,12.9817906 C11.9777302,12.5359633 10.8342622,12.3102877 9.69025521,12.3115003 C8.54261048,12.3102877 7.39968137,12.5359633 6.32506123,12.9817906 C5.2857408,13.4117195 4.31499881,14.0423986 3.47319507,14.8859539 C2.62963982,15.7282965 1.99896067,16.7020026 1.56768446,17.7372811 C1.12562971,18.8058383 0.89995409,19.944995 0.89995409,21.0841516 L0.89995409,21.1018014 L0.89995409,21.1194513 C0.89995409,22.2573953 1.12562971,23.3953394 1.57010963,24.4657828 C1.99896067,25.5040254 2.6308524,26.4746327 3.47319507,27.3169753 C4.31553773,28.159318 5.28681865,28.7924223 6.32506123,29.2224859 L25.2279205,36.6394667 L17.8097271,17.73782 Z"
-        transform="rotate(-90 33.583 31.004)"
-      />
-      <path
-        fill="#27AA7C"
-        d="M56.7715675,35.9690638 C57.8073849,35.5390002 58.7792048,34.908321 59.6228948,34.0647658 C60.4652374,33.2222884 61.0971292,32.2485823 61.5271928,31.2128996 L68.9452514,12.3112529 L50.0425269,19.7282337 C49.0042843,20.1581626 48.0324645,20.7912669 47.1906607,21.6336096 C46.348318,22.4759523 45.7164263,23.4460206 45.2875753,24.4849368 C44.8436343,25.5553803 44.6174197,26.6933243 44.6174197,27.8312684 L44.6174197,27.8489182 L44.6174197,27.8665681 C44.6174197,29.0057247 44.8429606,30.1448813 45.2851501,31.2141122 C45.7164263,32.2485823 46.3471055,33.222962 47.1906607,34.0653047 C48.0330034,34.90886 49.0030717,35.5395391 50.0425269,35.969468 C51.1178207,36.4147563 52.2600761,36.6409709 53.4077209,36.6397583 C54.5517278,36.6404319 55.6957348,36.4147563 56.7715675,35.9690638"
-        transform="rotate(-90 77.3 31.004)"
-      />
-      <path
-        fill="#F7685E"
-        d="M29.8931663 8.84084765C27.5063256 8.84084765 25.5710494 6.87735147 25.5710494 4.45326412 25.5710494 2.02905527 27.5063256.0655590919 29.8931663.0655590919 32.2801266.0655590919 34.2152831 2.02905527 34.2152831 4.45326412 34.2152831 6.87735147 32.2801266 8.84084765 29.8931663 8.84084765zM40.342342 8.84084765C37.9553817 8.84084765 36.0202252 6.87735147 36.0202252 4.45326412 36.0202252 2.02905527 37.9553817.0655590919 40.342342.0655590919 42.7291827.0655590919 44.6644589 2.02905527 44.6644589 4.45326412 44.6644589 6.87735147 42.7291827 8.84084765 40.342342 8.84084765z"
-        transform="translate(13.99 27.048)"
-      />
-    </g>
-  </svg>
-);
-
-const BeeLogo = ({ className, style }) => (
-  <svg
-    width="1320px"
-    width="930px"
-    viewBox="0 0 1320 930"
-    className={className}
-    style={style}
-  >
-    {/* Body */}
-    <path
-      transform="translate(450 200)"
-      fillRule="evenodd"
-      fill="rgb(238, 181, 13)"
-      fill="#E3B602"
-      d="M0.000,420.000 L420.000,420.000 L420.000,520.000 L0.000,520.000 L0.000,420.000 ZM0.000,220.000 L420.000,220.000 L420.000,310.000 L0.000,310.000 L0.000,220.000 ZM210.000,0.000 C289.776,0.000 359.161,44.484 394.706,110.000 L25.294,110.000 C60.839,44.484 130.224,0.000 210.000,0.000 ZM388.906,630.000 C351.938,689.998 285.645,730.000 210.000,730.000 C134.355,730.000 68.062,689.998 31.094,630.000 L388.906,630.000 Z"
-    />
-
-    {/* Left Wing */}
-    <path
-      transform="translate(0 280)"
-      fillRule="evenodd"
-      fill="rgb(38, 146, 60)"
-      fill="#27AA7C"
-      d="M460.000,-0.000 L331.000,386.000 L330.532,386.045 C307.370,452.394 244.249,500.000 170.000,500.000 C76.112,500.000 0.000,423.888 0.000,330.000 C0.000,262.255 39.629,203.772 96.967,176.451 L97.000,176.000 L460.000,-0.000 Z"
-    />
-
-    {/* Left Wing */}
-    <path
-      transform="translate(860 280)"
-      fillRule="evenodd"
-      fill="rgb(38, 146, 60)"
-      fill="#27AA7C"
-      d="M460.000,330.000 C460.000,423.888 383.888,500.000 290.000,500.000 C215.751,500.000 152.630,452.394 129.468,386.045 L129.000,386.000 L0.000,-0.000 L363.000,176.000 L363.033,176.451 C420.371,203.772 460.000,262.255 460.000,330.000 Z"
-    />
-
-    {/* Left Eye */}
-    <path
-      transform="translate(460 0)"
-      fillRule="evenodd"
-      fill="rgb(247, 175, 187)"
-      fill="#F7685E"
-      d="M95.000,0.000 C147.467,0.000 190.000,42.533 190.000,95.000 C190.000,147.467 147.467,190.000 95.000,190.000 C42.533,190.000 0.000,147.467 0.000,95.000 C0.000,42.533 42.533,0.000 95.000,0.000 Z"
-    />
-
-    <path
-      transform="translate(670 0)"
-      fillRule="evenodd"
-      fill="rgb(247, 175, 187)"
-      fill="#F7685E"
-      d="M95.000,0.000 C147.467,0.000 190.000,42.533 190.000,95.000 C190.000,147.467 147.467,190.000 95.000,190.000 C42.533,190.000 0.000,147.467 0.000,95.000 C0.000,42.533 42.533,0.000 95.000,0.000 Z"
-    />
-  </svg>
-);
-
-export default BeeLogo;
