+++ src/ui/client/src/components/BookingLayout/DestinationFragment/DestinationFragment.module.css
@@ -66,7 +66,6 @@
 .imageContainer {
   display: grid;
   grid-template-columns: repeat(3, 1fr);
-  /* margin-bottom: 2px; */
   margin: 0 24px 24px 24px;
   grid-gap: 2px;
   position: absolute;
@@ -77,7 +76,6 @@
 
 .image {
   width: 100%;
-  /* height: 136px; */
   height: 143px;
   object-fit: cover;
 }
