+++ src/ui/client/src/components/ErrorLayout/ErrorLayout.js
@@ -1,0 +1,3 @@
+import React from "react";
+const ErrorLayout = () => <>Error: Page does not exist</>;
+export default ErrorLayout;
