+++ src/ui/client/src/index.js
@@ -3,6 +3,7 @@ import ReactDOM from "react-dom";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 
+import "./ibmcloud.css";
 import "./index.css";
 import "./fonts.css";
